requirements.txt
fastapi
uvicorn
requests
tinkoff-invest
pydantic